refactor: create view folder and remove game folder
